Add serviceIntegrationEnabled in DiscoveryManager
serviceIntegrationEnabled variable is added in DiscoveryManager. If it
is true, services in a device are managed by one device instance
(default in the previous version). Otherwise, all services look like in
different devices (default in the current version).
The previous version showed all services in a device as a device item in
the DevicePickerList and automatically selected a service for a
capability requested by user.
It could give simple device list and usage scenario for a device with
multiple second screen services. However to support it more stably, more
careful design is needed under various cases:
- A service is not connected successfully.
- A service with high CapabilityPriorityLevel does not work.